tif-chatbot/requirements.txt
Flask==3.1.0
Flask_Cors==4.0.0
google_cloud_aiplatform==1.64.0
matplotlib==3.9.3
numpy==2.2.0
openai==1.57.3
pandas==2.2.3
prophet==1.1.6
protobuf==5.29.1
psycopg2_binary==2.9.10
pydantic==2.10.3
pyotp==2.9.0
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
qrcode==7.4.2
scikit_learn==1.6.0
seaborn==0.13.2
statsmodels==0.14.4
vertexai==1.64.0
minio
